Update options so that retry_delay is a size_t and retry_chance can be 0 to disable retries

+++ include/ares.h
@@ -336,7 +336,7 @@ struct apattern;
  */
 struct ares_server_failover_options {
   unsigned short retry_chance;
-  int            retry_delay;
+  size_t         retry_delay;
 };
 
 /* NOTE about the ares_options struct to users and developers.
